Made ShaderUniform into Uniform, abstracted, and set up exceptions if trying to call down to it. It is not MUCH simpler to create uniform bindings

+++ shaders/entityVertexShader.glsl
@@ -15,7 +15,7 @@ uniform mat4 projectionMatrix;
 uniform mat4 viewMatrix;
 uniform vec3 lightPosition[4];
 
-uniform float numberOfRows;
+uniform int numberOfRows;
 uniform vec2 offset;
 
 uniform float useFakeLighting;
